+++ packages/genesys-spark-components/src/components/stable/gux-table/gux-table-select-menu/gux-table-select-popover/gux-table-header-popover.tsx
@@ -11,8 +11,8 @@ import {
   Watch
 } from '@stencil/core';
 
-import { OnClickOutside } from '../../../../../utils/decorator/on-click-outside';
-import { onHiddenChange } from '../../../../../utils/dom/on-attribute-change';
+import { OnClickOutside } from '@utils/decorator/on-click-outside';
+import { onHiddenChange } from '@utils/dom/on-attribute-change';
 
 /**
  * @slot - popover content