+++ packages/genesys-spark-components/CHANGELOG.md
@@ -3636,7 +3636,7 @@ COMUI-343
 - **breadcrumbs component:** small style fixes ([0136ab2](http://inindca///commit/0136ab2a002f7b1bd842dee602c425d2a6ce48dd)), closes [#43](http://inindca///issues/43)
 - **breadcrumbs component:** update primary breadcrumbs. use GuxBreadcrumbAccent type for accent" ([8c15fff](http://inindca///commit/8c15fff10fef755ea8bbe30c982ac50eb0527f1f)), closes [#43](http://inindca///issues/43)
 - **breadcrumbs component:** update test ([b8ac1f4](http://inindca///commit/b8ac1f491fc269e575185adc001d1936b921cc3e)), closes [#43](http://inindca///issues/43)
-- **table component:** fix tests to use gux-table-beta selector ([4348007](http://inindca///commit/4348007d342bf1748e4596111411958f5138f92e))
+- **table component:** fix tests to use gux-table selector ([4348007](http://inindca///commit/4348007d342bf1748e4596111411958f5138f92e))
 
 ## [1.8.0](http://inindca///compare/v1.7.1...v1.8.0) (2020-09-28)